// frontend/src/api/model/githubCom4H1RZooraInternalDomainQAQuestionView.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export interface GithubCom4H1RZooraInternalDomainQAQuestionView {
  author_name?: string;
  closed_at?: string;
  created_at?: string;
  id?: string;
  model_id?: string;
  model_type?: string;
  status?: string;
  text?: string;
  user_id?: string;
  vote_count?: number;
  voted_by_me?: boolean;
}